Removed unnecessary new type. Refactored optIn_request
